code-refactoring: separate order page into components
