+++ client/src/constants/apiEndpoints.js
@@ -15,3 +15,7 @@ export const SIGNUP = (type) => `/auth/${type}/signup`;
 export const LOGIN = (type) => `/auth/${type}/login`;
 export const VERIFY = (type, token) => `/auth/${type}/verify/${token}`;
 
+
+// Orders
+export const ORDER_PRODUCT = (productId) => `/order/${productId}`;
+
