+++ server/routes/order.js
@@ -4,7 +4,7 @@ const orderController = require('../controllers/orderController');
 const verifyAccessToken = require('../middlewares/verifyAccessToken');
 
 // Add Order Item
-router.post("/", verifyAccessToken, orderController.addOrder);
+router.post("/:productId", verifyAccessToken, orderController.addOrder);
 
 // Retrieve Order Item according to Seller Id
 router.get("/", verifyAccessToken, orderController.showOrdersBySeller);
